dnsbl-ip.py: return ERR instead of BH for unresolvable destinations
BH caused Squid to display "permission denied" messages to the client,
which is confusing. ERR is considered to be safe here, as Squid won't
be able to establish a connection anyway, no matter wether the destination
is blacklisted or not, provided both Squid and dnsbl-ip.py use the same
DNS resolver.